[refactor] refactor ai plugins
