+++ shenyu-plugin/pom.xml
@@ -49,9 +49,7 @@
         <module>shenyu-plugin-proxy</module>
         <module>shenyu-plugin-security</module>
         <module>shenyu-plugin-fault-tolerance</module>
-        <module>shenyu-plugin-ai-proxy</module>
-        <module>shenyu-plugin-ai-prompt</module>
-        <module>shenyu-plugin-ai-token-limiter</module>
+        <module>shenyu-plugin-ai</module>
     </modules>
 
     <dependencies>
