+++ shenyu-plugin/shenyu-plugin-ai/pom.xml
@@ -1,0 +1,27 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<project xmlns="http://maven.apache.org/POM/4.0.0"
+         xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+    <modelVersion>4.0.0</modelVersion>
+    <packaging>pom</packaging>
+    <parent>
+        <groupId>org.apache.shenyu</groupId>
+        <artifactId>shenyu-plugin</artifactId>
+        <version>2.7.1-SNAPSHOT</version>
+    </parent>
+    
+    <artifactId>shenyu-plugin-ai</artifactId>
+    
+    <properties>
+        <maven.compiler.source>18</maven.compiler.source>
+        <maven.compiler.target>18</maven.compiler.target>
+        <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
+    </properties>
+
+    <modules>
+        <module>shenyu-plugin-ai-common</module>
+        <module>shenyu-plugin-ai-proxy</module>
+        <module>shenyu-plugin-ai-prompt</module>
+        <module>shenyu-plugin-ai-token-limiter</module>
+    </modules>
+</project>
