+++ shenyu-plugin/shenyu-plugin-ai/shenyu-plugin-ai-token-limiter/src/main/resources/META-INF/scripts/increment-token.lua
